Refactor

+++ HomeCenter.Actors.Tests/Helpers/Time.cs
@@ -4,7 +4,7 @@ namespace HomeCenter.Actors.Tests.Helpers
 {
     public static class Time
     {
-        public static long Tics(int miliseconds)
+        public static long Tics(double miliseconds)
         {
             return TimeSpan.FromMilliseconds(miliseconds).Ticks;
         }
